refactor: MIME, STATUS enum을 HttpUtil로 이동(Request, Response 공통 속성)
