+++ src/main/java/webserver/model/Response.java
@@ -1,60 +1,10 @@
 package webserver.model;
 
-import java.util.Arrays;
 import java.util.Map;
 
-import static http.HttpUtil.HEADER_HTTP_VERSION;
+import static http.HttpUtil.*;
 
 public class Response {
-    public enum MIME {
-        HTML("text/html"),
-        CSS("text/css"),
-        JS("text/javascript"),
-        ICO("image/vnd.microsoft.icon"),
-        PNG("image/png"),
-        JPG("image/jpeg"),
-        WOFF("application/x-font-woff"),
-        TTF("application/x-font-ttf");
-
-        private final String mime;
-        MIME(String mime) {
-            this.mime = mime;
-        }
-
-        public String getMime() {
-            return this.mime;
-        }
-
-        public static MIME getMimeByExtension(String extension) {
-            return Arrays.stream(MIME.values())
-                    .filter(mime -> extension.equalsIgnoreCase(mime.toString()))
-                    .findFirst()
-                    .orElse(null);
-        }
-    }
-    public enum STATUS {
-        OK(200, "OK"),
-        CREATED(201, "Created"),
-        TEMPORARY_MOVED(302, "Found"),
-        BAD_REQUEST(400, "Bad Request"),
-        UNAUTHORIZED(401, "Unauthorized"),
-        NOT_FOUND(404, "Not Found");
-
-        private final int statusCode;
-        private final String statusMessage;
-        STATUS(int statusCode, String statusMessage) {
-            this.statusCode = statusCode;
-            this.statusMessage = statusMessage;
-        }
-
-        public int getStatusCode() {
-            return statusCode;
-        }
-        public String getStatusMessage() {
-            return statusMessage;
-        }
-    }
-
     private final STATUS status;
     private final String version = HEADER_HTTP_VERSION;
     private final Map<String, String> headerMap;
